@@ -11,35 +11,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.13",
+    "@angular/animations": "^8.0.3",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "~7.2.13",
-    "@angular/compiler": "~7.2.13",
-    "@angular/core": "~7.2.13",
-    "@angular/forms": "~7.2.13",
+    "@angular/common": "~8.0.3",
+    "@angular/compiler": "~8.0.3",
+    "@angular/core": "~8.0.3",
+    "@angular/forms": "~8.0.3",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "~7.2.13",
-    "@angular/platform-browser-dynamic": "~7.2.13",
-    "@angular/router": "~7.2.13",
+    "@angular/platform-browser": "~8.0.3",
+    "@angular/platform-browser-dynamic": "~8.0.3",
+    "@angular/router": "~8.0.3",
     "chart.js": "^2.8.0",
     "codemirror": "^5.48.0",
     "core-js": "^2.6.2",
     "hammerjs": "^2.0.8",
     "highcharts": "^7.1.2",
     "ng2-file-upload": "^1.3.0",
-    "rxjs": "~6.4.0",
+    "rxjs": "~6.5.2",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "^7.2.13",
-    "@angular/language-service": "~7.2.13",
+    "@angular-devkit/build-angular": "~0.800.6",
+    "@angular/cli": "~8.0.6",
+    "@angular/compiler-cli": "^8.0.3",
+    "@angular/language-service": "~8.0.3",
     "@types/jasmine": "~3.3.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~10.12.18",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.1.0",
@@ -50,6 +50,6 @@
     "protractor": "~5.4.2",
     "ts-node": "~7.0.0",
     "tslint": "~5.12.1",
-    "typescript": "~3.2.4"
+    "typescript": "~3.4.5"
   }
 }
